fix: :bug: Back : To know if a job Information is already created or not

+++ app/controllers/advertissements.controller.js
@@ -74,51 +74,40 @@ const controller = {
     res.json(newAdvertisement);
   },
 
-  //! Create a job information
-  createJobInformation: async (req, res) => {
+  //! To know if the job information exist or not
+  isJobInformationExist: async (req, res) => {
     //Recup the userId from the token
     const token = req.headers.authorization?.replace("Bearer ", "");
     const user = loginService.getUser(token);
     const userId = user.id;
 
-    //Recup the advertId from the params & the jobInformation from the body
+    //Recup the advertId from the params
     const advertId = parseInt(req.params.advertId);
-    const { isSaved, isApplied } = req.body;
-    console.log(advertId, userId, isSaved, isApplied);
 
-    //Create the jobInformation
-    const newJobInformation = await datamapper.createJobInformation(
-      advertId,
-      userId,
-      isSaved,
-      isApplied
-    );
-    res.json(newJobInformation);
-  },
+    //Check if the job information exist
+    const jobInformation = await datamapper.getJobInformation(advertId, userId);
+    res.json(jobInformation);
 
-  //! Update an Job Information
+    if (jobInformation) {
+      //*Update the jobInformation
+      const updatedJobInformation = await datamapper.updateJobInformation(
+        advertId,
+        userId,
+        isSaved,
+        isApplied
+      );
 
-  updateJobInformation: async (req, res) => {
-    //Recup the userId from the token
-    const token = req.headers.authorization?.replace("Bearer ", "");
-    const user = loginService.getUser(token);
-    const userId = user.id;
-
-    //Recup the advertId from the params & the jobInformation from the body
-    const advertId = parseInt(req.params.advertId);
-    const { isSaved, isApplied } = req.body;
-
-    console.log(advertId, userId, isSaved, isApplied);
-
-    //Update the jobInformation
-    const updatedJobInformation = await datamapper.updateJobInformation(
-      advertId,
-      userId,
-      isSaved,
-      isApplied
-    );
-
-    res.json(updatedJobInformation);
+      res.json(updatedJobInformation);
+    } else {
+      //*Create the jobInformation
+      const newJobInformation = await datamapper.createJobInformation(
+        advertId,
+        userId,
+        isSaved,
+        isApplied
+      );
+      res.json(newJobInformation);
+    }
   },
 
   //! Show all saved advertisements
@@ -148,7 +137,38 @@ const controller = {
   //TODO: finish this delete route
   //! Delete an advertisement
   deleteAdvertisement: async (req, res) => {
+    //Recup the advertId
     const advertId = parseInt(req.params.advertId);
+
+    //Recup the userId from the token
+    const token = req.headers.authorization?.replace("Bearer ", "");
+    const user = loginService.getUser(token);
+    const userId = user.id;
+
+    //To know if the user is an admin
+    const isAdmin = await datamapper.isAdmin(userId);
+    if (!isAdmin) {
+      return res.status(403).send("You are not an admin");
+    }
+
+    //Check if the user is the admin of the company of the advertisement
+    const company = await datamapper.getCompanyByAdvertId(advertId);
+    if (company.userId !== userId) {
+      return res.status(403).send("You are not the admin of this company");
+    }
+
+    //Delete the advertisement
+    try {
+      const deletedAdvertisement = await datamapper.deleteAdvertisement(
+        advertId
+      );
+
+      res.json(deletedAdvertisement);
+    } catch (error) {
+      console.log(error);
+      res.status(500).send(error.message);
+    }
+
     const deletedAdvertisement = await datamapper.deleteAdvertisement(advertId);
     res.json(deletedAdvertisement);
   },
